refactor: renamed class imbalance algorithms (previous was 'sampling')
