+++ antropometria/__init__.py
@@ -3,6 +3,6 @@ from antropometria.main import main
 import antropometria.classifiers
 import antropometria.config
 import antropometria.feature_selectors
-import antropometria.sampling
+import antropometria.dataset_imbalance
 import antropometria.tests
 import antropometria.utils
