+++ antropometria/dataset_imbalance/__init__.py
@@ -1,0 +1,1 @@
+from antropometria.dataset_imbalance.ClassImbalanceReduction import ClassImbalanceReduction
